+++ packages/lab/src/collapsible/CollapsiblePanel.css
@@ -18,6 +18,6 @@
   visibility: visible;
 }
 
-.saltCollapsiblePanel[aria-hidden="true"] .saltCollapsiblePanel-inner {
+.saltCollapsiblePanel-inner {
   overflow: hidden;
 }
