+++ site/docs/components/navigation-item/usage.mdx
@@ -12,6 +12,10 @@ data:
 
 For further details on how to use the component to provide navigation, see the [navigation pattern](/salt/patterns/navigation/)
 
+### Routing libraries
+
+If you are using a routing library such as [React Router](https://reactrouter.com/en/main) or [Next.js](https://nextjs.org/docs/app/building-your-application/routing), you can use the `render` prop to render a link component from the routing library. This allows you to take advantage of the routing library's features, such as client-side navigation and route matching.
+
 ### When to use
 
 - When you have multiple levels of navigation in your application or website.
