+++ site/docs/components/vertical-navigation/index.mdx
@@ -1,7 +1,7 @@
 ---
 title: Vertical navigation
 data:
-  description: "`VerticalNavigation` allows users to navigate between distinct applications and website pages. VerticalNavigation supports two appearances: `indicator` and `bordered`."
+  description: "`VerticalNavigation` allows users to navigate between distinct applications and website pages."
   sourceCodeUrl: "https://github.com/jpmorganchase/salt-ds/blob/main/packages/lab/src/vertical-navigation"
   package:
     name: "@salt-ds/lab"
