+++ site/src/components/callout/Callout.tsx
@@ -2,13 +2,16 @@ import { Banner, BannerContent, H4, Text } from "@salt-ds/core";
 import type { ReactNode } from "react";
 import styles from "./Callout.module.css";
 
-type CalloutProps = { title: string; children: ReactNode };
+type CalloutProps = {
+  title: string;
+  children: ReactNode;
+};
 
-export const Callout = ({ title, children }: CalloutProps) => {
+export const Callout = ({ title, children, ...rest }: CalloutProps) => {
   return (
-    <Banner className={styles.root}>
+    <Banner className={styles.root} {...rest}>
       <BannerContent className={styles.content}>
-        <H4>{title}</H4>
+        {title && <H4>{title}</H4>}
         <Text className={styles.text}>{children}</Text>
       </BannerContent>
     </Banner>
