+++ site/src/examples/navigation-item/MockHistory.tsx
