+++ site/src/examples/navigation-item/WithBadge.tsx
@@ -1,43 +1,130 @@
-import { Badge, FlexLayout, NavigationItem } from "@salt-ds/core";
+import { Badge, NavigationItem, StackLayout } from "@salt-ds/core";
 import { type ReactElement, useState } from "react";
+import { Link, useLocation } from "react-router";
+import { MockHistory } from "./MockHistory";
 
-export const WithBadge = (): ReactElement => {
-  const [horizontalActive, setHorizontalActive] = useState(false);
+type NavItem = {
+  title: string;
+  href: string;
+  status?: string;
+  children?: NavItem[];
+};
 
-  const [verticalActive, setVerticalActive] = useState(false);
+const nested: NavItem[] = [
+  {
+    title: "Home",
+    href: "/",
+    status: "New",
+  },
+  {
+    title: "Products",
+    href: "/products",
+    children: [
+      {
+        title: "Widgets",
+        href: "/products/widgets",
+        children: [
+          { title: "Widget A", href: "/products/widgets/widget-a" },
+          { title: "Widget B", href: "/products/widgets/widget-b" },
+          { title: "Widget C", href: "/products/widgets/widget-c" },
+        ],
+      },
+      { title: "Gadgets", href: "/products/gadgets" },
+      { title: "Doodads", href: "/products/doodads" },
+    ],
+  },
+  {
+    title: "About Us",
+    href: "/about",
+    children: [
+      { title: "Our Story", href: "/about/story" },
+      { title: "Our Team", href: "/about/team" },
+      { title: "Press", href: "/about/press" },
+    ],
+  },
+  {
+    title: "Support",
+    href: "/support",
+  },
+  {
+    title: "Contact",
+    href: "/contact",
+  },
+];
 
-  const handleHorizontalActiveToggle = () => {
-    setHorizontalActive((current) => !current);
-  };
+function NavItem({ item, level = 0 }: { item: NavItem; level?: number }) {
+  const location = useLocation();
+  const [expanded, setExpanded] = useState(false);
 
-  const handleVerticalActiveToggle = () => {
-    setVerticalActive((current) => !current);
-  };
+  if (item.children && item.children.length > 0) {
+    return (
+      <li>
+        <NavigationItem
+          blurActive={location.pathname.startsWith(item.href) && !expanded}
+          orientation="vertical"
+          onExpand={() => {
+            setExpanded((old) => !old);
+          }}
+          parent
+          expanded={expanded}
+          level={level}
+        >
+          {item.title}
+        </NavigationItem>
+        {expanded && (
+          <StackLayout
+            as="ul"
+            gap="var(--salt-spacing-fixed-100)"
+            style={{
+              width: 250,
+              listStyle: "none",
+              paddingLeft: 0,
+            }}
+          >
+            {item.children?.map((item) => (
+              <NavItem key={item.href} item={item} level={level + 1} />
+            ))}
+          </StackLayout>
+        )}
+      </li>
+    );
+  }
 
   return (
-    <FlexLayout align="center">
+    <li>
       <NavigationItem
-        active={horizontalActive}
-        onClick={() => {
-          handleHorizontalActiveToggle();
-        }}
-        href="#"
-      >
-        Label
-        <Badge value="New" />
-      </NavigationItem>
-
-      <NavigationItem
-        active={verticalActive}
-        onClick={() => {
-          handleVerticalActiveToggle();
-        }}
-        href="#"
+        active={location.pathname === item.href}
+        href={item.href}
         orientation="vertical"
+        expanded={expanded}
+        render={<Link to={item.href} />}
+        level={level}
       >
-        Label
-        <Badge value="New" />
+        {item.title}
+        {item.status && <Badge value={item.status} />}
       </NavigationItem>
-    </FlexLayout>
+    </li>
+  );
+}
+
+export const WithBadge = (): ReactElement => {
+  return (
+    <MockHistory>
+      <nav>
+        <StackLayout
+          as="ul"
+          gap="var(--salt-spacing-fixed-100)"
+          style={{
+            width: 250,
+            listStyle: "none",
+            paddingLeft: 0,
+          }}
+        >
+          {nested.map((item) => (
+            <NavItem key={item.href} item={item} />
+          ))}
+        </StackLayout>
+      </nav>
+    </MockHistory>
   );
 };
