+++ site/src/examples/navigation-item/index.ts
@@ -1,11 +1,5 @@
-export * from "./BlurActive";
-export * from "./Default";
 export * from "./HorizontalGroup";
-export * from "./RenderElement";
-export * from "./RenderProp";
-export * from "./Vertical";
 export * from "./VerticalGroup";
 export * from "./VerticalNestedGroup";
 export * from "./WithBadge";
 export * from "./WithIcon";
-export * from "./WithNestedItems";
