+++ site/src/examples/vertical-navigation/MockHistory.tsx
@@ -1,0 +1,7 @@
+import type { ReactNode } from "react";
+import { MemoryRouter } from "react-router";
+
+export function MockHistory({ children }: { children: ReactNode }) {
+  // @ts-ignore
+  return <MemoryRouter>{children}</MemoryRouter>;
+}
