+++ site/src/examples/vertical-navigation/index.ts
@@ -1,4 +1,6 @@
+export * from "./Basic";
 export * from "./CollapsibleSubmenu";
 export * from "./MultiActionItem";
-export * from "./Nested";
+export * from "./Submenu";
+export * from "./WithBadges";
 export * from "./WithIcons";
